+++ src/process.rs
@@ -3,7 +3,6 @@
 use crate::argument::*;
 use crate::block::*;
 use crate::inst::*;
-use crate::module::ModuleContext;
 use crate::seq_body::*;
 use crate::ty::*;
 use crate::unit::*;
@@ -116,24 +115,16 @@ impl Value for Process {
 }
 
 pub struct ProcessContext<'tctx> {
-    module: &'tctx ModuleContext<'tctx>,
     process: &'tctx Process,
 }
 
 impl<'tctx> ProcessContext<'tctx> {
-    pub fn new(module: &'tctx ModuleContext, process: &'tctx Process) -> ProcessContext<'tctx> {
-        ProcessContext {
-            module: module,
-            process: process,
-        }
+    pub fn new(process: &'tctx Process) -> ProcessContext<'tctx> {
+        ProcessContext { process: process }
     }
 }
 
 impl<'tctx> Context for ProcessContext<'tctx> {
-    fn parent(&self) -> Option<&Context> {
-        Some(self.module.as_context())
-    }
-
     fn try_value(&self, value: &ValueRef) -> Option<&Value> {
         match *value {
             ValueRef::Inst(id) => Some(self.inst(id)),