+++ CHANGELOG.md
@@ -17,6 +17,7 @@ parse_module` to the `assembly` module.
 ### Removed
 - Remove `write` and `write_string` from the `assembly` module.
 - Remove `visit` module.
+- Remove `Module` in favor of `ir::Module`.
 
 ### Fixed
 - Fix emission of time constants.